fix: server simplify
