+++ server/src/db.js
@@ -1,5 +1,4 @@
 const knex = require('knex')
-const createTasksTables = require("./tasks/ddl")
 
 async function connect(options) {
   const db = knex({
@@ -9,6 +8,19 @@ async function connect(options) {
 
   await createTasksTables(db);
   return db
+}
+
+async function createTasksTables(db) {
+  const tasksExists = await db.schema.hasTable('tasks')
+  if (!tasksExists) {
+    await db.schema.createTable('tasks', function(table) {
+        table.string('title')
+        table.string('description')
+        table.integer('version') // Required for conflict resolution
+        table.increments('id')
+        table.string('status')
+    })
+  }
 }
 
 module.exports = connect
