+++ server/src/index.js
@@ -1,74 +1,48 @@
 const express = require('express')
 const cors = require('cors')
+const unifiedPushSenderClient = require('unifiedpush-node-sender')
 const { VoyagerServer } = require('@aerogear/voyager-server')
 const { KeycloakSecurityService } = require('@aerogear/voyager-keycloak')
+const { createSubscriptionServer } = require('@aerogear/voyager-subscriptions')
+
 const config = require('./config/config')
 const connect = require('./db')
-const { createSubscriptionServer } = require('@aerogear/voyager-subscriptions')
-const { appTypeDefs, appResolvers } = require('./schema')
-const agSender = require("unifiedpush-node-sender")
+const { typeDefs, resolvers } = require('./schema')
 
 process.on('unhandledRejection', error => {
-  console.error('unhandledRejection', error.message);
-  process.exit(1);
-});
-
-let keycloakService = null
-let pushClient = null
-
-// if a keycloak config is present we create
-// a keycloak service which will be passed into
-// ApolloVoyagerServer
-if (config.keycloakConfig) {
-  keycloakService = new KeycloakSecurityService(config.keycloakConfig)
-}
-
-if (config.pushConfig) {
-  let pushService = agSender(config.pushConfig)
-  pushClient = pushService.then((client) => {
-    return client;
-  })
-}
+  console.error(error.message, error.stack)
+  process.exit(1)
+})
 
 async function start() {
 
   const app = express()
-
   app.use(cors())
+  app.use('/', express.static('website'))
+  app.get('/health', (req, res) => res.sendStatus(200))
 
-  if (keycloakService) {
+  let keycloakService = null
+  let pushClient = null
+
+  if (config.keycloakConfig) {
+    keycloakService = new KeycloakSecurityService(config.keycloakConfig)
     keycloakService.applyAuthMiddleware(app)
   }
 
-  app.use('/', express.static('website'))
+  if (config.pushConfig) {
+    pushClient =  await unifiedPushSenderClient(config.pushConfig)
+  }
 
-  app.get('/health', (req, res) => res.sendStatus(200))
-
-  // connect to db
-  const client = await connect(config.db);
+  const db = await connect(config.db)
 
   const apolloConfig = {
-    typeDefs: appTypeDefs,
-    resolvers: appResolvers,
+    typeDefs,
+    resolvers,
     playground: config.playgroundConfig,
     introspection: true,
-    context: async ({
-      req
-    }) => {
-      // pass request + db ref into context for each resolver
-      return {
-        req: req,
-        db: client,
-        pushClient: await pushClient
-      }
-    },
-    uploads: {
-      // Limits here should be stricter than config for surrounding
-      // infrastructure such as Nginx so errors can be handled elegantly by
-      // graphql-upload:
-      // https://github.com/jaydenseric/graphql-upload#type-uploadoptions
-      maxFileSize: 10000000, // 10 MB
-      maxFiles: 5
+    context: {
+      db,
+      pushClient
     }
   }
 
@@ -81,10 +55,6 @@ async function start() {
   apolloServer.applyMiddleware({ app })
 
   const server = app.listen(config.port, () => {
-    console.log(`\n    ***********************************************************
-    🎮 Ionic PWA application available at http://localhost:${config.port}
-    🚀 GraphQL Playground is available at http://localhost:${config.port}/graphql
-    ***********************************************************`)
     createSubscriptionServer({
       securityService: keycloakService,
       schema: apolloServer.schema
@@ -95,4 +65,9 @@ async function start() {
   })
 }
 
-start()
+start().then(() => {
+  console.log(`\n    ***********************************************************
+    🎮 Ionic PWA application available at http://localhost:${config.port}
+    🚀 GraphQL Playground is available at http://localhost:${config.port}/graphql
+    ***********************************************************`)
+})
