--- server/src/tasks/ddl.js
@@ -1,20 +1,0 @@
-
-
-async function createTasksTables(db) {
-    const tasksExists = await db.schema.hasTable('tasks')
-    if (!tasksExists) {
-        await db.schema.createTable('tasks', function(table) {
-            table.string('title')
-            table.string('description')
-            // Required for conflict resolution
-            table.integer('version')
-            table.increments('id')
-            table.string('status')
-        })
-    }
-}
-
-module.exports = createTasksTables
-
-
-
