--- server/src/tasks/index.js
@@ -1,8 +1,0 @@
-
-const { taskTypeDefs, taskResolvers } = require('./schema')
-const { subscriptionTypeDefs, subscriptionResolvers } = require('./subscriptions')
-
-module.exports = {
-    taskTypeDefs, taskResolvers,
-    subscriptionTypeDefs, subscriptionResolvers
-}
